Added scrollview around dynamic buttons for loadable characters. Still has issue where size of scroll rect changes.
To do:

Either add in "notes" tab in screen hub or see about making loadable character data more specific (show class and level, check if xml file is one for the application, etc). May also start looking into letting users add an avatar picture for their characters (we may be able to also do this for like the backgrounds in the screen hub to give a little user customization, but that can be last-minute).